backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary>=2.9.10
django-environ==0.11.2
Pillow>=10.2.0
celery==5.3.4
redis==5.0.1
django-ratelimit==4.1.0
django-filter==23.5
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
faker==20.1.0

